v 0.8.5 compat, melee text, continued
* Created this continuation mod.
* Quasimorph 0.8.5 compatibility.
* Added "M" for melee.
