--- QM-DisplayMovementSpeed.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34728.123
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QM-DisplayMovementSpeed", "QM-DisplayMovementSpeed\QM-DisplayMovementSpeed.csproj", "{76F5A6F2-93DA-499D-8BC9-052B4F1E1BD0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{76F5A6F2-93DA-499D-8BC9-052B4F1E1BD0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{76F5A6F2-93DA-499D-8BC9-052B4F1E1BD0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{76F5A6F2-93DA-499D-8BC9-052B4F1E1BD0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{76F5A6F2-93DA-499D-8BC9-052B4F1E1BD0}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A78DD824-9BF1-4243-B74E-22963DBB9673}
-	EndGlobalSection
-EndGlobal
