--- QM-DisplayMovementSpeed/HideTextMesh.cs
@@ -1,49 +1,0 @@
-﻿using MGSC;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using UnityEngine;
-using static System.Net.Mime.MediaTypeNames;
-using TMPro;
-
-namespace QM_DisplayMovementSpeed
-{
-    public class HideTextMesh : MonoBehaviour
-    {
-        public TextMeshPro textMesh;
-        public SpriteRenderer spriteRenderer;
-        public Monster monster;
-        void Start()
-        {
-            SpriteRenderer[] spriteRenderers = gameObject.GetComponentsInChildren<SpriteRenderer>();
-            foreach(SpriteRenderer s in spriteRenderers)
-            {
-                if (s.name == "shadow")
-                {
-                    spriteRenderer = s;
-                    break;
-                }
-            }
-            if (spriteRenderer == null)
-            {
-                spriteRenderer = gameObject.transform.GetChild(gameObject.transform.childCount - 1).GetComponent<SpriteRenderer>();
-            }
-            textMesh = gameObject.GetComponentInChildren<TextMeshPro>();
-
-
-        }
-
-        void FixedUpdate ()
-        {
-            
-            if (textMesh.renderer == null) return;
-
-            if (Plugin.show)
-                textMesh.renderer.enabled = spriteRenderer.enabled;
-            else
-                textMesh.renderer.enabled = false;
-        }
-    }
-}
