--- QM-DisplayMovementSpeed/QM-DisplayMovementSpeed.csproj
@@ -1,63 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <OutputType>Library</OutputType>
-    <RootNamespace>QM_DisplayMovementSpeed</RootNamespace>
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-  </PropertyGroup>
-	<PropertyGroup>
-		<GAME_NAME>Quasimorph</GAME_NAME>
-		<WorkshopId>3282796258</WorkshopId>
-	</PropertyGroup>
-    <PropertyGroup>
-	    <GAME_PATH Condition="!Exists('$(GAME_PATH)')">$([MSBuild]::GetRegistryValueFromView('HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\Steam App $(GAME_STEAM_ID)', 'InstallLocation', null, RegistryView.Registry64, RegistryView.Registry32))</GAME_PATH>
-	    <_SteamLibraryPath>$([MSBuild]::GetRegistryValueFromView('HKEY_CURRENT_USER\SOFTWARE\Valve\Steam', 'SteamPath', null, RegistryView.Registry32))</_SteamLibraryPath>
-	    <GAME_PATH Condition="!Exists('$(GAME_PATH)') AND '$(_SteamLibraryPath)' != ''">$(_SteamLibraryPath)\steamapps\common\$(GAME_NAME)</GAME_PATH>
-	    <GAME_PATH Condition="!Exists('$(GAME_PATH)')">C:\Program Files\Steam\steamapps\common\$(GAME_NAME)</GAME_PATH>
-	    <GAME_PATH Condition="!Exists('$(GAME_PATH)')">C:\Program Files (x86)\Steam\steamapps\common\$(GAME_NAME)</GAME_PATH>
-		<WorkshopPath Condition="Exists('$(GAME_PATH)')">$(GAME_PATH)\..\..\workshop\content\2059170\$(WorkshopId)</WorkshopPath>
-    </PropertyGroup>
-	<ItemGroup>
-    <Reference Include="Assembly-CSharp" publicize="true">
-      <HintPath>$(GAME_PATH)\Quasimorph_Data\Managed\Assembly-CSharp.dll</HintPath>
-    </Reference>
-    <Reference Include="Unity.TextMeshPro, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <SpecificVersion>False</SpecificVersion>
-      <HintPath>$(GAME_PATH)\Quasimorph_Data\Managed\Unity.TextMeshPro.dll</HintPath>
-    </Reference>
-		<Reference Include="UnityEngine.UI, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-			<SpecificVersion>False</SpecificVersion>
-			<HintPath>$(GAME_PATH)\Quasimorph_Data\Managed\UnityEngine.UI.dll</HintPath>
-		</Reference>
-		<Reference Include="UnityEngine">
-      <HintPath>$(GAME_PATH)\Quasimorph_Data\Managed\UnityEngine.dll</HintPath>
-    </Reference>
-    <Reference Include="UnityEngine.CoreModule">
-      <HintPath>$(GAME_PATH)\Quasimorph_Data\Managed\UnityEngine.CoreModule.dll</HintPath>
-    </Reference>
-    <Reference Include="UnityEngine.TextRenderingModule, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <SpecificVersion>False</SpecificVersion>
-      <HintPath>$(GAME_PATH)\Quasimorph_Data\Managed\UnityEngine.TextRenderingModule.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="BepInEx.AssemblyPublicizer.MSBuild" Version="0.4.2">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="lib.harmony" Version="2.3.3" />
-    <PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="8.0.0" />
-    <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="System.Buffers" Version="4.5.1" />
-    <PackageReference Include="System.Data.DataSetExtensions" Version="4.5.0" />
-    <PackageReference Include="System.Memory" Version="4.5.5" />
-    <PackageReference Include="System.Numerics.Vectors" Version="4.5.0" />
-    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="6.0.0" />
-    <PackageReference Include="System.Text.Encodings.Web" Version="8.0.0" />
-    <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.5.4" />
-    <PackageReference Include="System.ValueTuple" Version="4.5.0" />
-  </ItemGroup>
-  <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-    <Exec Command="if &quot;$(WorkshopPath)&quot;==&quot;&quot; (echo workshop path not found &amp;&amp; exit /b)&#xD;&#xA;&#xD;&#xA;&#xD;&#xA;xcopy /f /y  &quot;$(ProjectDir)$(OutDir)$(TargetName).dll&quot; &quot;$(WorkshopPath)\&quot;&#xD;&#xA;xcopy /f /y  &quot;$(ProjectDir)$(OutDir)$(TargetName).pdb&quot; &quot;$(WorkshopPath)\&quot;&#xD;&#xA;xcopy /f /y  &quot;$(ProjectDir)$(OutDir)0Harmony.dll&quot; &quot;$(WorkshopPath)\&quot;&#xD;&#xA;&#xD;&#xA;if &quot;$(Configuration)&quot;==&quot;Release&quot; (del &quot;$(WorkshopPath)\*.pdb&quot;)&#xD;&#xA;&#xD;&#xA;&#xD;&#xA;" />
-  </Target>
-</Project>
