+++ SteamReadMe.txt
@@ -1,0 +1,52 @@
+[h1]Display Enemy Movement Speed - Continued...[/h1]
+
+
+[h1]Functionality[/h1]
+
+For the people constantly right clicking on enemies to not get owned by 3AP melee goons.
+Shows the amount of action points an enemy can spend. If the enemy is melee only, an 'M' will also be displayed.
+
+Toggleable with the comma key, rebindable with a config file.
+
+[h1]Continued Notes[/h1]
+
+This is a continuation of GitHub user [url=https://github.com/jamsge]jamsge's[/url] fantastic mod of the same name.
+If the original author comes back, this mod will be removed in favor of their mod listing.
+
+[h2][b]Important![/b][/h2]
+
+Make sure you are not subscribed to the original Display Enemy Movement Speed mod, or this mod will not work correctly.
+
+[h2]Config[/h2]
+
+To rebind the comma key toggle, go to:
+[i]%UserProfile%\AppData\LocalLow\Magnum Scriptum Ltd\Quasimorph_ModConfigs\QM_DisplayMovementSpeedContinued.json[/i]
+and change "Comma" [url=https://docs.unity3d.com/ScriptReference/KeyCode.html]with a Unity Keycode of your choice (scroll down to the Properties section)[/url].
+
+[h2]Source[/h2]
+
+https://github.com/NBKRedSpy/QM-DisplayMovementSpeedContinued
+
+[h2]Credits[/h2]
+
+[url=https://www.textstudio.com/]"Continued" font created with textstudio.com[/url]
+
+[h2]FAQ[/h2]
+
+[i][b]I ran into an bug, what do I do?[/b][/i]
+If you run into issues, please let me know through comment or forum post with details so I can try to make fixes.
+
+[i][b]Can you make XYZ feature?[/b][/i]
+I can't guarantee anything, but please feel free to leave a comment or forum post with suggestions.
+
+*** I can't find the config file! What do I do? ***
+You need to run the game at least once for a config file to generate
+
+[h1]Change Log[/h1]
+
+[h2]1.1.0[/h2]
+[list]
+[*]Created this continuation mod.
+[*]Quasimorph 0.8.5 compatibility.
+[*]Added "M" for melee.
+[/list]
